% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agricultural_chargeoff_rates_by_quarter}
\alias{agricultural_chargeoff_rates_by_quarter}
\title{Agricultural Chargeoff Rates by Quarter}
\format{A \code{data.frame} with 100 rows and 3 columns indicating the chargeoff rates for agricultural loans for each quarter.}
\source{
\url{https://www.federalreserve.gov/releases/chargeoff/chgallnsa.htm}
}
\usage{
agricultural_chargeoff_rates_by_quarter
}
\description{
This data set contains twenty-five years (1991 through 2015) of quarterly charge-off rates for agricultural loans as reported by the Federal Reserve. This is the full dataset corresponding to the selected rows displayed in Table 2.5 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(agricultural_chargeoff_rates_by_quarter)
summary(agricultural_chargeoff_rates_by_quarter)
}
\keyword{datasets}
