% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airline_arrivals}
\alias{airline_arrivals}
\title{Airline Arrivals}
\format{A \code{data.frame} with 13 rows and 13 columns indicating the percentages and ranks for each airline.}
\source{
U.S. Department of Transportation (2016)
}
\usage{
airline_arrivals
}
\description{
This data set contains on-time arrival records for U.S. flight carriers in for all four quarters of the year 2015, the month of December 2015, and the entire year 2015. This data is displayed in Table 2.9 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(airline_arrivals)
summary(airline_arrivals)
}
\keyword{datasets}
