% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{american_league_salary_2014}
\alias{american_league_salary_2014}
\title{American League Salary 2014}
\format{A \code{data.frame} with 447 rows and 3 columns corresponding to the name, salary, and team for each player.}
\source{
Petchesky (2014)
}
\usage{
american_league_salary_2014
}
\description{
This data set contains the 2014 salaries (as of March 26, 2014) for all baseball players in the American League.
}
\examples{
data(american_league_salary_2014)
summary(american_league_salary_2014)
}
\keyword{datasets}
