% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arion_subfuscus}
\alias{arion_subfuscus}
\title{Arion Subfuscus}
\format{A \code{data.frame} with 10 rows and 2 columns indicating the AI for the woodland and waste sites.}
\source{
Whelan (1982)
}
\usage{
arion_subfuscus
}
\description{
This data set contains the Acceptability Indices (AI) for Arion Subfuscus from woodland and waste sites with the toxic woodland plant Allium Ursinum (wild garlic) as test gel as reported by Whelan (1982). This data is displayed in Table 9.3 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(arion_subfuscus)
summary(arion_subfuscus)
}
\keyword{datasets}
