% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{average_HDL_levels}
\alias{average_HDL_levels}
\title{Average HDL Levels}
\format{A \code{vector} with 12 observations corresponding to the average HDL measurements for each woman.}
\source{
Kerr (1983)
}
\usage{
average_HDL_levels
}
\description{
This data set contains the averages of two HDL measurements taken 24 hours apart for each of twelve women who  participated in Kerr (1983). This data is displayed in Table 7.1 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(average_HDL_levels)
summary(average_HDL_levels)
}
\keyword{datasets}
