% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beer_head}
\alias{beer_head}
\title{Beer Head}
\format{A \code{list} containing 2 \code{data.frame}s (one for each of the two production lots). Each \code{data.frame} has 20 rows and 3 columns corresponding to the attributes for each bottle.}
\source{
Ault et al. (1967)
}
\usage{
beer_head
}
\description{
This data set contains the height of the initial head formation upon pouring, the percentage adhesion of the head, and the percentage collapse of the head 4 min after pouring for 20 bottles selected from two different production lots of beer as reported by Ault et al. (1967).
}
\examples{
data(beer_head)
summary(beer_head)
}
\keyword{datasets}
