% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chargeoff_rates}
\alias{chargeoff_rates}
\title{Chargeoff Rates}
\format{A \code{data.frame} with 100 rows and 9 columns indicating the chargeoff rates for each of the eight loan types for each quarter.}
\source{
\url{https://www.federalreserve.gov/releases/chargeoff/chgallnsa.htm}
}
\usage{
chargeoff_rates
}
\description{
This data set contains twenty-five years (1991 through 2015) of quarterly charge-off rates for eight different types of loans as reported by the Federal Reserve. This data is displayed in Table 2.4 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(chargeoff_rates)
summary(chargeoff_rates)
}
\keyword{datasets}
