% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{college_rankings_2012}
\alias{college_rankings_2012}
\title{College Rankings 2012}
\format{A \code{data.frame} with 7793 rows and 9 columns indicating various scorecard metrics for each school.}
\source{
United States Department of Education (2016)
}
\usage{
college_rankings_2012
}
\description{
This data set contains a subset of the College Scorecard Data reported by the U.S. Department of Education.
}
\examples{
data(college_rankings_2012)
summary(college_rankings_2012)
}
\keyword{datasets}
