% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{desimipramine}
\alias{desimipramine}
\title{Desimipramine}
\format{A \code{data.frame} with 12 rows and 2 columns indicating the pre-treatment and post-treatment weights (in kg) of each subject.}
\source{
Brambilla et al. (1985)
}
\usage{
desimipramine
}
\description{
This data set contains the body weight of 12 subjects with anorexia nervosa before and after treatment with the medication desimipramine as reported by Brambilla et al. (1985). This data is displayed in Table 8.3 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(desimipramine)
summary(desimipramine)
}
\keyword{datasets}
