% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diamonds_carats_color_cost}
\alias{diamonds_carats_color_cost}
\title{Diamonds Carats Color Cost}
\format{A \code{data.frame} with 308 rows and 5 columns indicating the attributes of each diamond.}
\source{
Chu (2001)
}
\usage{
diamonds_carats_color_cost
}
\description{
This data set contains weight in carats, color purity, grade of clarity, certification body, and value in Singapore dollars for 308 round diamond stones from an advertisement in the February 18, 2000 edition of Singapore's Business Times as discussed in Chu (2001).
}
\examples{
data(diamonds_carats_color_cost)
summary(diamonds_carats_color_cost)
}
\keyword{datasets}
