% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emissions}
\alias{emissions}
\title{Emissions}
\format{A \code{data.frame} with 25 rows and 5 columns indicating the amount of the four pollutants for each year.}
\source{
U.S. Environmental Protection Agency (2016)
}
\usage{
emissions
}
\description{
This data set contains annual emissions data (in million metric tons of carbon dioxide equivalents) for four common air pollutants from 1990 to 2014 in the United States. This data is displayed in Table 2.3 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(emissions)
summary(emissions)
}
\keyword{datasets}
