% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{engineering_drawing_hours}
\alias{engineering_drawing_hours}
\title{Engineering Drawing Hours}
\format{A \code{vector} with 96 observations corresponding to the hours for each piece of machinery/mechanical device.}
\source{
Ohio-based company
}
\usage{
engineering_drawing_hours
}
\description{
This data set contains the total engineering drawing hours contributing to the cost of pieces of machinery/mechanical devices for a large Ohio-based company that is displayed in Table 1.3 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(engineering_drawing_hours)
summary(engineering_drawing_hours)
}
\keyword{datasets}
