% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{female_amerindians}
\alias{female_amerindians}
\title{Female Amerindians}
\format{A \code{vector} with 20 observations corresponding to the stature (in centimeters) for each female Amerindian.}
\source{
Sciulli and Carlisle (1975)
}
\usage{
female_amerindians
}
\description{
This data set contains the stature of 20 female Amerindians from the Turner site in Cincinnati, Ohio, representing the Ohio Hopewell period (200-400 AD) studied by Sciulli and Carlisle (1975). This data is displayed in Table 7.8 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(female_amerindians)
summary(female_amerindians)
}
\keyword{datasets}
