% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmr_white_tailed_deer}
\alias{fmr_white_tailed_deer}
\title{FMR White-Tailed Deer}
\format{A \code{list} with 3 \code{vector}s, each containing FMR (kcal/kg/day) for the deer observed in the corresponding time period.}
\source{
Silver et al. (1969)
}
\usage{
fmr_white_tailed_deer
}
\description{
This data set contains data on the fasting metabolic rate (FMR) of white-tailed deer for different two-month periods of the year. This data is displayed in Table 12.9 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(fmr_white_tailed_deer)
summary(fmr_white_tailed_deer)
}
\keyword{datasets}
