% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender_roles}
\alias{gender_roles}
\title{Gender Roles}
\format{A \code{data.frame} with 7 rows and 2 columns indicating the counts for each gender and role category.}
\source{
Vigorito and Curry (1998)
}
\usage{
gender_roles
}
\description{
This data set contains counts of gender 7685 individuals for which gender could be identified into the seven role categories. This data was used by Vigorito and Curry (1998) to investigate whether there is any relationship between gender and the roles that individuals play in magazine ad illustrations. This data is displayed in Table 10.12 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(gender_roles)
summary(gender_roles)
}
\keyword{datasets}
