% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goggled_green_turtles}
\alias{goggled_green_turtles}
\title{Goggled Green Turtles}
\format{A \code{data.frame} with 18 rows and 2 columns indicating the number of circles in a two-minute period for the nasal field goggles and for the harlequin goggles.}
\source{
Mrosovsky and Shettleworth (1974)
}
\usage{
goggled_green_turtles
}
\description{
This data set contains the number of times that 36 turtle hatchlings split into two groups 'circled', in an attempt to determine whether light had an effect on the orientation and sea-finding ability of such hatchlings as reported by Mrosovsky and Shettleworth (1974). This data is displayed in Table 9.6 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(goggled_green_turtles)
summary(goggled_green_turtles)
}
\keyword{datasets}
