% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health_care_by_affiliation}
\alias{health_care_by_affiliation}
\title{Health Care by Affiliation}
\format{A \code{data.frame} with 2 rows and 2 columns indicating the counts for each statement and each party.}
\source{
Princeton Survey Research Associates of Princeton, New Jersey (1998)
}
\usage{
health_care_by_affiliation
}
\description{
This data set contains counts of respondents' preferences between two statements about healthcare grouped by political party from the report "Generation to Generation: American Values about Taking Care of Each Other" (1998). This data is displayed in Table 10.20 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(health_care_by_affiliation)
summary(health_care_by_affiliation)
}
\keyword{datasets}
