% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homes_prices}
\alias{homes_prices}
\title{Homes Prices}
\format{A \code{data.frame} with 100 rows and 6 columns indicating the attributes of each house.}
\source{
Woodard and Leone (2008)
}
\usage{
homes_prices
}
\description{
This data set contains the information about house prices for a random sample of 100 properties in Wake County, North Carolina, as collected by Woodard and Leone (2008).
}
\examples{
data(homes_prices)
summary(homes_prices)
}
\keyword{datasets}
