% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_lot_sizes}
\alias{house_lot_sizes}
\title{House Lot Sizes}
\format{A \code{data.frame} with 100 rows and 8 columns indicating the attributes of each house.}
\source{
Woodard and Leone (2008)
}
\usage{
house_lot_sizes
}
\description{
This data set contains the information about house and lot sizes for a random sample of 100 properties in Wake County, North Carolina, as collected by Woodard and Leone (2008).
}
\examples{
data(house_lot_sizes)
summary(house_lot_sizes)
}
\keyword{datasets}
