\name{IIS-package}
\alias{IIS-package}
\alias{IIS}
\docType{package}
\title{
Datasets and functions to accompany Wolfe and Schneider - Intuitive Introductory Statistics  (2017)
}
\description{
This package contains the R datasets and functions referenced in the text and exercises of Wolfe and Schneider - Intuitive Introductory Statistics (2017)
}
\author{
Grant Schneider

Maintainer: Grant Schneider <gschneidz@gmail.com>

}
\references{
Wolfe and Schneider - Intuitive Introductory Statistics (2017)
}
\keyword{Intuitive}
\keyword{Introductory}
\keyword{Wolfe}
\keyword{Schneider}
