% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infant_walking}
\alias{infant_walking}
\title{Infant Walking}
\format{A \code{data.frame} with 6 rows and 2 columns indicating the age (in months) at which infants first walked for the exercise and no-exercise groups.}
\source{
Zelazo et al. (1972)
}
\usage{
infant_walking
}
\description{
This data set contains the ages at which 12 infants were reported to have started walking by their mothers. The infants were randomly split into "exercise" and "no-exercise" groups" as part of the study conducted by Zelazo et al. (1972). This data is displayed in Table 9.8 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(infant_walking)
summary(infant_walking)
}
\keyword{datasets}
