% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interstitial_lengths}
\alias{interstitial_lengths}
\title{Interstitial Lengths}
\format{A \code{data.frame} with 12 rows and 2 columns corresponding to 12 observations each of blue spruce and white pine trees.}
\source{
Kayle (1984)
}
\usage{
interstitial_lengths
}
\description{
This data set contains the averages (mm) of interstitial lengths (distances between midpoints) of ten pairs of secondary branches for each of twelve blue spruce and twelve white pine trees. This data is displayed in Table 1.24 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(interstitial_lengths)
summary(interstitial_lengths)
}
\keyword{datasets}
