% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kentucky_derby_2012}
\alias{kentucky_derby_2012}
\title{Kentucky Derby 2012}
\format{A \code{data.frame} with 23 rows and 6 columns indicating various statistics for each year.}
\source{
\url{https://kentuckyderby.ag/}
}
\usage{
kentucky_derby_2012
}
\description{
This data set contains race statistics for the Kentucky Derby for each year from 1990 to 2012. This data is displayed in Table 2.7 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(kentucky_derby_2012)
summary(kentucky_derby_2012)
}
\keyword{datasets}
