% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meniscal_repairs_load_at_failure}
\alias{meniscal_repairs_load_at_failure}
\title{Meniscal Repairs Load at Failure}
\format{A \code{list} with 3 \code{vector}s, each containing the load (in Newtons) at failure of the 6 menisci assigned to the corresponding group.}
\source{
Borden et al. (2003)
}
\usage{
meniscal_repairs_load_at_failure
}
\description{
This data set contains the load at failure for 18 cadaveric menisci repaired by one of three techniques: the FasT-Fix Meniscal Repair Suture System (FasT-Fix), the use of biodegradable Meniscus Arrows (MA), and the Vertical Mattress Sutures (VMS) approach. This data is displayed in Table 12.1 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(meniscal_repairs_load_at_failure)
summary(meniscal_repairs_load_at_failure)
}
\keyword{datasets}
