% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mother_smoking_age}
\alias{mother_smoking_age}
\title{Mother Smoking Age}
\format{A \code{data.frame} with 45 rows and 4 columns indicating the age group, number of births, and percentage of smoking mothers for each year.}
\source{
Centers for Disease Control and Prevention (CDC) "WONDER Online Database" (2016)
}
\usage{
mother_smoking_age
}
\description{
This data set contains the percentages of mothers in the United States who smoked during pregnancies from 2010 to 2014 for various age categories. This data is displayed in Table 2.14 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(mother_smoking_age)
summary(mother_smoking_age)
}
\keyword{datasets}
