% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mother_smoking_education}
\alias{mother_smoking_education}
\title{Mother Smoking Education}
\format{A \code{data.frame} with 40 rows and 4 columns indicating the education, births, and percentage of smoking mothers for each year.}
\source{
Centers for Disease Control and Prevention (CDC) "WONDER Online Database" (2016)
}
\usage{
mother_smoking_education
}
\description{
This data set contains the percentages of mothers in the United States from 2010 to 2014 who smoked during pregnancy for various categories of educational level. A subset of this data is displayed in Table 2.13 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(mother_smoking_education)
summary(mother_smoking_education)
}
\keyword{datasets}
