% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mother_smoking_education_1989_1993}
\alias{mother_smoking_education_1989_1993}
\title{Mother Smoking Education 1989-1993}
\format{A \code{data.frame} with 25 rows and 3 columns indicating the education and percentage of smoking mothers for each year.}
\source{
Centers for Disease Control and Prevention, National Center for Health Statistics (1995)
}
\usage{
mother_smoking_education_1989_1993
}
\description{
This data set contains the percentages of mothers in the United States from 1989 to 1993 who smoked during pregnancy for various categories of educational level. This data is displayed in Table 2.15 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(mother_smoking_education_1989_1993)
summary(mother_smoking_education_1989_1993)
}
\keyword{datasets}
