% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mother_smoking_education_2010}
\alias{mother_smoking_education_2010}
\title{Mother Smoking Education 2010}
\format{A \code{data.frame} with 8 rows and 2 columns indicating the percentage of smoking mothers for each educational level in the year 2010.}
\source{
Centers for Disease Control and Prevention (CDC) "WONDER Online Database" (2016)
}
\usage{
mother_smoking_education_2010
}
\description{
This data set contains the percentages of mothers in the United States in 2010 who smoked during pregnancy for various categories of educational level. This is a subset of the mother_smoking_education dataset.
}
\examples{
data(mother_smoking_education_2010)
summary(mother_smoking_education_2010)
}
\keyword{datasets}
