% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motor_vehicle_death_rate_2012}
\alias{motor_vehicle_death_rate_2012}
\title{Motor Vehicle Death Rate 2012}
\format{A \code{data.frame} with 50 rows and 3 columns corresponding to the name, fatality rate, and region for each state.}
\source{
National Highway Traffic Safety Administration (2013)
}
\usage{
motor_vehicle_death_rate_2012
}
\description{
This data set contains the total motor-vehicle fatalities per 100 million vehicle miles traveled in the 2012 calendar year that is displayed in Table 1.16 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(motor_vehicle_death_rate_2012)
summary(motor_vehicle_death_rate_2012)
}
\keyword{datasets}
