% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movie_facts}
\alias{movie_facts}
\title{Movie Facts}
\format{A \code{data.frame} with 100 rows and 6 columns corresponding to the title, release year, length (minutes), number of cast members, rating, and number of lines of description for each movie.}
\source{
Moore (2006)
}
\usage{
movie_facts
}
\description{
This data set contains a random sample of 100 movies from the 1996 edition of The Movie and Video Guide prepared annually by Leonard Maltin as selected by Moore (2006).
}
\examples{
data(movie_facts)
summary(movie_facts)
}
\keyword{datasets}
