% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{national_league_salary_2014}
\alias{national_league_salary_2014}
\title{National League Salary 2014}
\format{A \code{data.frame} with 437 rows and 3 columns corresponding to the name, salary, and team for each player.}
\source{
Petchesky (2014)
}
\usage{
national_league_salary_2014
}
\description{
This data set contains the 2014 salaries (as of March 26, 2014) for all baseball players in the National League.
}
\examples{
data(national_league_salary_2014)
summary(national_league_salary_2014)
}
\keyword{datasets}
