% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nba_2015_2016}
\alias{nba_2015_2016}
\title{NBA 2015-2016}
\format{A \code{data.frame} with 30 rows and 19 columns indicating various performance statistics for each team.}
\source{
\url{https://www.nba.com/stats/teams/traditional?Season=2015-16}
}
\usage{
nba_2015_2016
}
\description{
This data set contains NBA teams' performance in the 2015-2016 season as provided by the NBA's official website.
}
\examples{
data(nba_2015_2016)
summary(nba_2015_2016)
}
\keyword{datasets}
