% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osu_math_salaries_2015}
\alias{osu_math_salaries_2015}
\title{OSU Math Salaries 2015}
\format{A \code{vector} with 226 observations corresponding to each reported salary.}
\source{
Cleveland.com (2016)
}
\usage{
osu_math_salaries_2015
}
\description{
This data set contains 2015 salaries of faculty and staff in the Mathematics Department at The Ohio State University. This data is displayed in Table 3.1 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(osu_math_salaries_2015)
summary(osu_math_salaries_2015)
}
\keyword{datasets}
