% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pennies_age}
\alias{pennies_age}
\title{Pennies' Age}
\format{A \code{vector} with 374 observations corresponding to the ages of the pennies.}
\source{
Classroom Experiment
}
\usage{
pennies_age
}
\description{
This data set contains the ages (based on minting dates) of 374 United States pennies collected by an introductory class of 25 students.
}
\examples{
data(pennies_age)
summary(pennies_age)
}
\keyword{datasets}
