% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{percentage_hatched_eggs}
\alias{percentage_hatched_eggs}
\title{Percentage Hatched Eggs}
\format{A \code{data.frame} with 9 rows and 2 columns indicating the percentage of hatched eggs for the two nest types for each year.}
\source{
Clawson et al. (1979)
}
\usage{
percentage_hatched_eggs
}
\description{
This data set contains the percentages of eggs that eventually hatched for dump (communal) nests and normal nests each year from 1966 through 1974 as reported by Clawson et al. (1979). This data is displayed in Table 8.1 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(percentage_hatched_eggs)
summary(percentage_hatched_eggs)
}
\keyword{datasets}
