% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pew_science_survey_data_by_age_group}
\alias{pew_science_survey_data_by_age_group}
\title{Pew Science Survey Data By Age Group}
\format{A \code{data.frame} with 5 rows and 4 columns indicating the proportion of respondents answering "yes" to each of the four questions indicated in Exercise 2.3.4. of Wolfe and Schneider - Intuitive Introductory Statistics by age group.}
\source{
Pew Internet & American Life Project, Science Issues (2014)
}
\usage{
pew_science_survey_data_by_age_group
}
\description{
This data set contains the percentages of people who answered "yes" to various questions about scientific topics in a survey conducted by the Pew Research Center grouped by age group. This data is displayed in Table 2.18 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(pew_science_survey_data_by_age_group)
summary(pew_science_survey_data_by_age_group)
}
\keyword{datasets}
