% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pew_science_survey_data_by_party}
\alias{pew_science_survey_data_by_party}
\title{Pew Science Survey Data By Party}
\format{A \code{data.frame} with 3 rows and 4 columns indicating the proportion of respondents answering "yes" to each of the four questions indicated in Exercise 2.3.4. of Wolfe and Schneider - Intuitive Introductory Statistics by party.}
\source{
Pew Internet & American Life Project, Science Issues (2014)
}
\usage{
pew_science_survey_data_by_party
}
\description{
This data set contains the percentages of people who answered "yes" to various questions about scientific topics in a survey conducted by the Pew Research Center grouped by political affiliation. This data is displayed in Table 2.17 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(pew_science_survey_data_by_party)
summary(pew_science_survey_data_by_party)
}
\keyword{datasets}
