% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pines_1997}
\alias{pines_1997}
\title{Pines 1997}
\format{A \code{data.frame} with 1000 rows and 5 columns corresponding to variables described in Table 2.2 of Intuitive Introductory Statistics for each tree.}
\source{
Kenyon Center for Environmental Study (1997)
}
\usage{
pines_1997
}
\description{
This data set contains a subset of the data collected by biology students at Kenyon College to determine which factors cause pines in the Kenyon Center for Environmental Study to vary in growth rates. A description of the dataset is displayed in Table 2.2 and Example 3.4 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(pines_1997)
summary(pines_1997)
}
\keyword{datasets}
