% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmn_migration}
\alias{pmn_migration}
\title{PMN Migration}
\format{A \code{vector} with 8 observations corresponding to the PMN migration for each patient.}
\source{
Perez et al. (1983)
}
\usage{
pmn_migration
}
\description{
This data set contains the polymorphonuclear leukocytes (PMN) migration in the sera of the 8 patients with acute pancreatitis measured by Perez et al. (1983). This data is displayed in Table 7.4 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(pmn_migration)
summary(pmn_migration)
}
\keyword{datasets}
