% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_estimates_2015}
\alias{population_estimates_2015}
\title{Population Estimates 2015}
\format{A \code{data.frame} with 50 rows and 17 columns indicating each metric over the four years for each state.}
\source{
United States Census Bureau (2016)
}
\usage{
population_estimates_2015
}
\description{
This data set contains population estimates, birth rates (per 1000 population), and net migration (per 1000 population) for each year 2011 through 2015 at the state level as reported by the U.S. Census Bureau.
}
\examples{
data(population_estimates_2015)
summary(population_estimates_2015)
}
\keyword{datasets}
