% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presidential_election_polls}
\alias{presidential_election_polls}
\title{Presidential Election Polls}
\format{A \code{data.frame} with 46 rows and 5 columns corresponding to the year, name, final polling estimate, actual election results, and deviation between polling and actual results for each candidate.}
\source{
Gallup (2015)
}
\usage{
presidential_election_polls
}
\description{
This data set contains the results of the final pre-election Gallup Voter Opinion Poll and the eventual election results for every presidential election from 1936 through 2012.
}
\examples{
data(presidential_election_polls)
summary(presidential_election_polls)
}
\keyword{datasets}
