% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportion_for_profit_hospitals}
\alias{proportion_for_profit_hospitals}
\title{Proportion For-Profit Hospitals}
\format{A \code{data.frame} with 20 rows and 3 columns indicating the region and proportion of for-profit hospitals for each state.}
\source{
American Hospital Association (2012)
}
\usage{
proportion_for_profit_hospitals
}
\description{
This data set contains data on the proportion of a state's hospitals that are for-profit for a subset of states from each of four regions of the country. This data is displayed in Table 12.5 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(proportion_for_profit_hospitals)
summary(proportion_for_profit_hospitals)
}
\keyword{datasets}
