% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{q2_q4_agricultural_chargeoff_rates}
\alias{q2_q4_agricultural_chargeoff_rates}
\title{Q2/Q4 Agricultural Chargeoff Rates}
\format{A \code{data.frame} with 50 rows and 3 columns indicating the chargeoff rates for agricultural loans for each quarter.}
\source{
\url{https://www.federalreserve.gov/releases/chargeoff/chgallnsa.htm}
}
\usage{
q2_q4_agricultural_chargeoff_rates
}
\description{
This data set contains twenty-five years (1991 through 2015) of Q2 and Q4 charge-off rates for agricultural loans as reported by the Federal Reserve. This is a subset of the agricultural_chargeoff_rates_by_quarter dataset.
}
\examples{
data(q2_q4_agricultural_chargeoff_rates)
summary(q2_q4_agricultural_chargeoff_rates)
}
\keyword{datasets}
