% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reading_habits_2011}
\alias{reading_habits_2011}
\title{Reading Habits 2011}
\format{A \code{data.frame} with 2986 rows and 7 columns indicating various pieces of demographic information and reading habits for each study participant.}
\source{
Pew Internet & American Life Project: The rise of e-reading (2016)
}
\usage{
reading_habits_2011
}
\description{
This data set contains data from the 2011 Pew Research Center report titled "The Rise of E-reading" intended to investigate how reading and e-reading habits vary by demographic and socio-economic categories. A subset of this data is displayed in Table 3.4 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(reading_habits_2011)
summary(reading_habits_2011)
}
\keyword{datasets}
