% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school_report_cards_2014}
\alias{school_report_cards_2014}
\title{School Report Cards 2014}
\format{A \code{data.frame} with 484 rows and 8 columns indicating various performance statistics for each high school.}
\source{
New York City Department of Education (2016)
}
\usage{
school_report_cards_2014
}
\description{
This data set contains 2014 school performance data, or "School Quality Snapshots", for high schools in New York City. A subset of this data is displayed in Table 3.2 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(school_report_cards_2014)
summary(school_report_cards_2014)
}
\keyword{datasets}
