% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sheep_weight}
\alias{sheep_weight}
\title{Sheep Weight}
\format{A \code{data.frame} with 20 rows and 2 columns indicating the weight (in kg) for the mother and the lamb of each pair.}
\source{
Ozturk et al. (2005)
}
\usage{
sheep_weight
}
\description{
This data set contains the mother's mating weight and her lamb offspring's weight at age 7 months for a subset of twenty ewe-lamb pairs from the Research Farm at Ataturk University, Erzurum, Turkey. This data is displayed in Table 11.6 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(sheep_weight)
summary(sheep_weight)
}
\keyword{datasets}
