% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_cdi}
\alias{state_cdi}
\title{State CDI}
\format{A \code{data.frame} with 16 rows and 3 columns indicating the indicator and the percentage of the popoulation having such an indicator for each state.}
\source{
Centers for Disease Control and Prevention (2016)
}
\usage{
state_cdi
}
\description{
This data set contains a sample of "chronic disease indicators", as reported by the CDC's Division of Population Health, for the four states California, Michigan, Ohio, and West Virginia. This data is displayed in Table 2.16 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(state_cdi)
summary(state_cdi)
}
\keyword{datasets}
