% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_poverty_levels_2013}
\alias{state_poverty_levels_2013}
\title{State Poverty Levels 2013}
\format{A \code{data.frame} with 50 rows and 3 columns corresponding to the name, total population, and number of people at or below the poverty line in each state.}
\source{
United States Census Bureau: Population Division (2014) & American Community Surveys (2014)
}
\usage{
state_poverty_levels_2013
}
\description{
This data set contains the number of individuals in each state whose annual income was below the established poverty level in 2013 at a state-by-state level. This data is displayed in Table 1.4 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(state_poverty_levels_2013)
summary(state_poverty_levels_2013)
}
\keyword{datasets}
