% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiaa_cref}
\alias{tiaa_cref}
\title{TIAA CREF}
\format{A \code{data.frame} with 302 rows and 8 columns indicating the unit values for each of the seven annuities for each day.}
\source{
\url{https://www.tiaa-cref.org}
}
\usage{
tiaa_cref
}
\description{
This data set contains the historical unit values from January 2, 2015 to March 15, 2016 for seven TIAA and CREF variable annuities. This is the full dataset corresponding to the selected rows displayed in Table 2.6 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(tiaa_cref)
summary(tiaa_cref)
}
\keyword{datasets}
