% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traffic_accidents}
\alias{traffic_accidents}
\title{Traffic Accident Data}
\format{A \code{data.frame} with 9 rows and 2 columns. The rows correspond to various types of accidents and the two columns correspond to the years 1945 and 1985, respectively.}
\source{
National Safety Council (1996)
}
\usage{
traffic_accidents
}
\description{
This data set contains traffic accident data from 1949 and 1985 that is displayed in Table 1.2 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(traffic_accidents)
summary(traffic_accidents)
}
\keyword{datasets}
