% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weekly_salaries}
\alias{weekly_salaries}
\title{Weekly Salaries}
\format{A \code{data.frame} with 44 rows and 6 columns corresponding to male and female earnings for each of the three occupation types for each quarter.}
\source{
U.S. Department of Labor, Bureau of Labor Statistics (2016)
}
\usage{
weekly_salaries
}
\description{
This data set contains the median weekly earnings of male and female employees in service occupations, sales and office occupations, and construction and extraction occupations for each quarter from 2005 to 2015. This data is displayed in Table 2.1 of Wolfe and Schneider - Intuitive Introductory Statistics.
}
\examples{
data(weekly_salaries)
summary(weekly_salaries)
}
\keyword{datasets}
