% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight_of_Euros}
\alias{weight_of_Euros}
\title{Weight of Euros}
\format{A \code{data.frame} with 2000 rows and 3 columns corresponding to the ID, weight, and batch/package number for each coin.}
\source{
Shkedy et al. (2006)
}
\usage{
weight_of_Euros
}
\description{
This data set contains the weight of 2000 coins from 8 separate packages of 250 Euros each as weighed and reported by Shkedy et al. (2006).
}
\examples{
data(weight_of_Euros)
summary(weight_of_Euros)
}
\keyword{datasets}
